requirements.txt
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
basemap-data==1.3.2
blessed==1.20.0
blinker==1.6.2
Cartopy==0.22.0
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
colorama==0.4.6
conda-pack==0.7.1
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.3
cycler==0.11.0
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.9.0
Flask==2.3.3
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
geos==0.2.3
gitdb==4.0.10
GitPython==3.1.32
gpustat==1.1.1
h5py==3.9.0
hickle==5.0.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.4
lightning-utilities==0.9.0
lit==15.0.7
llvmlite==0.41.0
lpips==0.1.4
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
model-index==0.1.11
mpmath==1.2.1
multidict==6.0.4
netCDF4==1.6.4
networkx==3.0
ninja==1.11.1
numba==0.58.0
numpy==1.24.1
nvidia-ml-py==12.535.133
opencv-python==4.8.0.74
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.19
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.3.0
platformdirs==3.9.1
protobuf==4.24.2
psutil==5.9.5
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.2
rich==13.4.2
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.4.0
scipy==1.11.1
sentry-sdk==1.30.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.30
timm==0.9.2
tokenizers==0.14.1
tomli==2.0.1
torch-scatter==2.1.2
torch_geometric==2.4.0
torchmetrics==1.0.1
tqdm==4.65.0
transformers==4.34.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.9
wcwidth==0.2.12
Werkzeug==2.3.7
xarray==2023.7.0
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
